requirements_optional.txt
django-allauth
django-messages
